[WIP] Add a Test Runner
The test runner will open an 'inferior' BSQL process and interact with
it by writing to its standard input and reading from their standard
output. It writes to the inferior process the lines from the test script
that start with "> " and verifies that the what the inferior process
writes to the standard output matches the content of the lines that
start with "< ".

A test script looks like

  > insert 1 foo@gmail.com 22
  < Ok.
  > .exit

See the file tests/smoke for a test script example.